refactor: streamline Clang, GCC, and MSVC command parsing by removing unused parameters and consolidating model building
